+++ lib/types.ts
@@ -1,2 +1,2 @@
-export type SOBData = Record<string, any>
-export type SOBRecord = SOBData & {id: string}
+export type SOBData = Record<string, any>;
+export type SOBRecord = SOBData & { id: string };