+++ releases.md
@@ -1,4 +1,11 @@
 # Release updates
+
 #### 1.1.0
-* TypeScript implementation
-* Change body of DELETE /:db/:table/delete
+
+- TypeScript implementation
+- Change body of DELETE /:db/:table/delete
+
+#### 1.2.0
+
+- Add statistics recording using env SHOW_STATISTICS
+- Add POST /:db/:table/update/batch